CS62 Data Structures Class Assignment - Darwin
This assignment helped me to: 
Gain practice writing large, multi-class programs.
Learn the importance of modular decomposition and information hiding. (The entire program is broken down into a series of classes that can be developed and tested independently, without revealing representational details.)
Have fun with a problem that is algorithmically interesting in its own right!